Move Livewire components from ./tournament to ./tournaments
