+++ resources/views/livewire/tournaments/team/player-actions.blade.php
